Day 6 tests, completed.
Well, after a loooong time I finally completed day 6 from the 2016 tests. This was really tricky one for me, and I don't even know why but it took longer than it should mainly because I kept failing to understand the question multiple times but as I grasped it, piece of cake (lol). On to the next